Add FDM data file extraction code & docs
This is a very rough tool to extract data from output files provided by
FDM who did the send for Council Tax bills in March 2020.

By comparing the data this tool extracts we can make educated guesses as
to who we should and shouldn't be sending council tax bills to and who
might be eligible for a reduction or discount.

The python tool was written in a trial and error way, and is pretty
brittle. It loops through the file and when matching lines to regex
or strings are found they are added to an array to be written as a csv.

Tests and saving values to a dict rather than appending to an array would
better, but difficult given we just read the file line by line and
have a lot of variables keeping track of where we are at any given time,
but could work. Hopefully this script will never need to be used ever
now I have most of the data out the bill file data.