Add FDM Data exemption reason extractor
This script takes the exemption reasons from FDM data into a CSV and SQL
files so that we can append the exemption reason to the existing FDM
data.

*INPUT DATA*: This was proving to be a challenge, so we manually updated the
input files to feature "class [A-Z]" on its own line. So for example
using a regex Find and replace on `class [A-Z]` we output the class and
classification letter on a new line. So the old data file would inclue
(newlines are literal here):

```
because it falls into the category of exempt class N which is OCCUPIED - ALL
OCCUPIERS ARE STUDENTS.
```
The curated data for extraction should be:
```
because it falls into the category of exempt
class N
 which is OCCUPIED - ALL
OCCUPIERS ARE STUDENTS.
```