Add example to add reason for excluding accounts
We will initially get data good enough to create bills by excluding
accounts that we think there is a good chance we can not send an
accurate bill, for example when people get discounts.

To keep track of which accounts we are excluding we will match each
account from the main dataset (aws_academy_recovered_october) with a
reason for exclusion in a join table.

This query is an example of adding all accounts that got a discount to
the exclusions list.