Add Death Report PDF extractor script
This is really brittle and not ideal, but we needed a way to get out
lots of data from a set of PDF files which include report meta data
as well as the data we need.

The script takes the pdf, gets all the text and then with a load of
hand crafted replace and regular expressions on extracted text, outputs
a pipe separated CSV. Pipe was chosen as the inconsistent commas in the
addresses were causing csv import to Postgres to not match columns.
Wrapping the addresses in quotes was tried but imports still picked up
the commas as an extra column.

I think there might be some superfluous steps in the code, but as it's
working, refactoring later is wise if the code needs to be reworked for
a live system, but for a proof of concept it serves our purpose.