Add initial Flask app
This includes a bootstrapped Flask app for working with Python and our
AWS Postgres DB.

Using the same connection through an SSH tunnel it uses psycopg2
postgres package to connect and query the DB.

Flask was chosen as it's lightweight, will solve a few templating
problems and gives us a basic web interface to work with, which will
be useful in previewing and manipulating council tax data. We can run
our pre-prepared queries and have access to them in Python for
formatting and templating the data output for the fulfilment company.

Some possible problems include handling the size of the data returned,
as it will be significant, the basic Flask app is not production ready
and it might be difficult.