Restructure app a bit to be more sensible
This restructures the app and splits things out a little to make the structure a bit more sensible.
It's by no means meeting the usual Flask structure but it's a bit more sensible now.

Splitting everything out from __init__.py would be a sensible next step and adding tests around the
little code there is to grab an account details using account_number.